requirements.txt
streamlit>=1.20.0
opencv-python>=4.7.0
Pillow>=9.0.0
numpy>=1.23.0
scipy>=1.8.0
scikit-learn>=1.1.0
plotly>=5.0.0
pandas>=1.4.0
# Optional / dev
kaleido>=0.2.1  # optional: plotly PNG export
pytest>=7.0.0    # optional: unit tests
